Move getRawTransaction into helpers
